Only call emptyDirSync on "dist" if it exists
As a side-effect, emptyDirSync will create the folder if it does not exist.  Small annoyance for users who are not using the default "dist" directory for build artifacts.

+++ app/lib/artifacts.js
@@ -65,6 +65,8 @@ module.exports.cleanAll = function () {
   fse.removeSync(folder)
   log(`Cleaned build artifact: "${folder}"`)
 
-  fse.emptyDirSync(appPaths.resolve.app('dist'))
-  log(`Emptied dist folder`)
+  if (fs.existsSync('dist')) {
+    fse.emptyDirSync(appPaths.resolve.app('dist'))
+    log(`Emptied dist folder`)
+  }
 }
